Deal with subtitles file download errors
Sometimes `download_file` will raise an error due to a transient server problem.

In this case, the subtitle's `_sid` field will not be set, causing a
crash at the next execution which can't be recovered from.

This fix ignores an unset `_sid` field to allow the user to continue
cycling through the subtitles list.