Make OSD delays configurable
The *long* delay is for messages that indicate a potentially long action
is ongoing (i.e. Downloading subtitles). The goal is to keep the message
on the screen for as long as the action is going.

Those *long* messages are always followed by a *short* confirmation
message which will override the previous *long* one when the action is
finished.

Old behavior can be achieved by setting the delays to `-1`